requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]